seperate tab into design tab
